Add tests, allow for variants from multiple sources.
